feat(mobile): immersive chat — replace the floating header with menu FABs
Drop the floating chat header bar in favour of a modern immersive panel: the
chat fills the screen and a few small glass FABs float over the corners —
menu (with the pending-actions badge) on the left to open the nav drawer,
rename + session actions on the right. The container is pointerEvents="box-none"
so the chat scrolls/touches through the gap between clusters. Other screens keep
their top app bar (TopBar). Removes FloatingChatHeader; updates the layout test.
